Moving helpers functions to helpers file
